+++ biopipen/scripts/vcf/TruvariConsistency.sh
@@ -1,6 +1,0 @@
-vcfs={{in.vcfs | join: " " | quote}}
-outfile={{out.outfile | quote}}
-truvari={{envs.truvari | quote}}
-
-$truvari consistency \
-    $vcfs > "$outfile"